# VM Loop optimization