--- a/sandbox/MRubyD.ConsoleApp/ruby/test.rb
+++ b/sandbox/MRubyD.ConsoleApp/ruby/test.rb
@@ -1,12 +1,4 @@
-def fb
-    n = 0
-        Proc.new do
-        n += 1
-        case
-            when n % 15 == 0
-            else n
-        end
-    end
+a=0
+while a < 100000
+  a+=1
 end
-
-fb.call