```rb
a=0
while a < 100000
  a+=1
end
```
```
| Method      | Mean       | Error     | StdDev    | Allocated |
|------------ |-----------:|----------:|----------:|----------:|
| MRubyD      | 2,778.3 us | 253.65 us | 167.77 us |     394 B |
| MRubyD (PR) | 1,035.2 us |   5.28 us |   3.49 us |     392 B |
| MRubyNative |   955.4 us |   6.42 us |   3.82 us |         - |

```