@hadashiA 2025-03-27
> Is this a mistake?

Yeah, it looks like your version is correct ><